refactoring - moved TagFactory to Common
(maybe ProcessorBase should be in Core)
