+++ WebScrap.Common/Tags/Creators/ClosingTagCreator.cs
@@ -1,9 +1,9 @@
 using WebScrap.Common.Tags;
 using WebScrap.Common.Contracts;
 
-namespace WebScrap.Tags.Creators;
+namespace WebScrap.Common.Tags.Creators;
 
-internal class ClosingTagCreator : ITagCreator
+public class ClosingTagCreator : ITagCreator
 {
     public TagBase Create(ReadOnlySpan<char> tag)
         => new ClosingTag(tag.ToString());
