+++ WebScrap.Common/Tags/Creators/OpeningTagCreator.cs
@@ -1,23 +1,25 @@
 using WebScrap.Common.Tags;
+using WebScrap.Common.Tags.Extractors;
 using WebScrap.Common.Contracts;
-using static WebScrap.Tags.Attributes.AttributeExtractor;
 
-namespace WebScrap.Tags.Creators;
+namespace WebScrap.Common.Tags.Creators;
 
-internal class OpeningTagCreator : ITagCreator
+public class OpeningTagCreator : ITagCreator
 {
     public TagBase Create(ReadOnlySpan<char> tagContent)
     {
         var index = tagContent.IndexOf(' ');
         if (index == -1)
         {
-            return new OpeningTag(tagContent.ToString(), EmptyAttributes);
+            return new OpeningTag(
+                tagContent.ToString(), 
+                AttributeExtractor.EmptyAttributes);
         }
         var name = tagContent[..index];
         tagContent = tagContent[tagContent.IndexOf(' ')..][1..];
         
         return new OpeningTag(
             name.ToString(), 
-            GetAttributes(tagContent));
+            AttributeExtractor.GetAttributes(tagContent));
     }
 }
